+++ packages/native/src/components/view/masonry-grid/masonry-grid-item.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { View } from 'react-native'
+import type { MasonryGridItemProps } from './masonry-grid-item.type'
+
+export const MasonryGridItem: React.FC<MasonryGridItemProps> = ({
+  children,
+  style,
+}) => {
+  return <View style={style}>{children}</View>
+}
+
+MasonryGridItem.displayName = 'MasonryGridItem'